requirements.txt
aiohttp==3.7.3
APScheduler==3.7.0
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
discord-ext-menus==1.0.0a27+g6f2b873
discord.py==1.6.0
idna==3.1
multidict==5.1.0
pytz==2021.1
six==1.15.0
typing-extensions==3.7.4.3
tzlocal==2.1
yarl==1.6.3
